@@ -1,41 +1,33 @@
 {
   "name": "pes-fe",
   "version": "0.1.0",
-  "proxy": "http://localhost:8080",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.13.0",
-    "@emotion/styled": "^11.13.0",
-    "@fortawesome/fontawesome-svg-core": "^6.5.1",
-    "@fortawesome/free-solid-svg-icons": "^6.5.1",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "@lezer/highlight": "^1.2.0",
-    "@mui/icons-material": "^5.16.7",
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^6.0.2",
     "@mui/joy": "^5.0.0-beta.48",
-    "@mui/material": "^5.16.7",
-    "@mui/styled-engine-sc": "^6.0.0-alpha.18",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^29.5.12",
-    "ace-builds": "^1.32.6",
+    "@mui/material": "^6.0.2",
+    "@reduxjs/toolkit": "^2.3.0",
+    "ace-builds": "^1.36.2",
     "axios": "^1.6.5",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-ace": "^10.1.0",
-    "react-cookie": "^7.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.0.1",
-    "react-input-mask": "^2.0.4",
+    "react-redux": "^9.1.2",
     "react-router-dom": "^6.21.1",
-    "react-scripts": "^5.0.1",
+    "react-spinners": "^0.14.1",
+    "redux": "^5.0.1",
+    "redux-persist": "^6.0.0",
     "styled-components": "^6.1.12",
-    "web-vitals": "^2.1.4"
+    "webpack-bundle-analyzer": "^4.10.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "webpack serve --open",
+    "build": "webpack",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "eject": "echo \"Error: no eject script specified\" && exit 1"
   },
   "eslintConfig": {
     "extends": [
@@ -56,17 +48,22 @@
     ]
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@types/node": "^22.5.0",
+    "@babel/core": "^7.25.7",
+    "@babel/plugin-transform-private-property-in-object": "^7.25.7",
+    "@babel/preset-env": "^7.25.7",
+    "@babel/preset-react": "^7.25.7",
+    "@babel/preset-typescript": "^7.25.7",
     "@types/react": "^18.3.4",
     "@types/react-dom": "^18.3.0",
-    "@types/react-input-mask": "^3.0.5",
-    "eslint-config-react-app": "^7.0.1",
+    "babel-loader": "^9.1.3",
+    "compression-webpack-plugin": "^11.1.0",
+    "css-loader": "^7.1.2",
     "html-webpack-plugin": "^5.6.0",
-    "ts-loader": "^9.5.1",
-    "typescript": "^4.9.5",
+    "image-webpack-loader": "^8.1.0",
+    "mini-css-extract-plugin": "^2.9.2",
+    "style-loader": "^4.0.0",
     "webpack": "^5.94.0",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.4"
+    "webpack-dev-server": "^5.1.0"
   }
 }
