@@ -1,31 +1,20 @@
-// import React from "react";
-// import ReactDOM from "react-dom/client";
-// import App from "./App";
-// import { BrowserRouter } from "react-router-dom";
-// import "styles/css/Font.css";
-
-// const rootElement = document.getElementById("root") as HTMLElement;
-// const root = ReactDOM.createRoot(rootElement);
-
-// root.render(
-//   <BrowserRouter>
-//     {/* <React.StrictMode> */}
-//     <App />
-//     {/* </React.StrictMode> */}
-//   </BrowserRouter>
-// );
-
-
-
-// 기존코드
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "./App";
 import { BrowserRouter } from "react-router-dom";
-import "styles/css/Font.css";
+import { Provider } from "react-redux";
+import { PersistGate } from "redux-persist/integration/react";
+import store, { persistor } from "stores/store";
+import Loading from "components/common/Loading";
 
-ReactDOM.render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
-  document.getElementById("root") 
+const rootElement = document.getElementById("root") as HTMLElement;
+const root = ReactDOM.createRoot(rootElement);
+
+root.render(
+  <Provider store={store}>
+    <PersistGate loading={<Loading />} persistor={persistor}> {/* Redux 상태가 복원되기 전까지 보여줄 로딩 화면 */}
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </PersistGate>
+  </Provider>
 );
