@@ -1,8 +1,5 @@
 import axios from "axios";
-import { Member } from "model/Store";
-
-// 회원 데이터 상태관리 훅 setMemDate의 타입
-type SetMemDate = React.Dispatch<React.SetStateAction<Member[]>>;
+import { Member, SetMemDate } from "model/userType";
 
 // 회원목록 api호출
 const getUsersList = async (setMemDate: SetMemDate): Promise<void> => {
