@@ -1,9 +1,5 @@
 import axios from "axios";
-import React from "react";
-import { MyFeedback } from "model/Store";
-
-// myFeedback 상태관리 HOOK setMyFeedback 타입
-type SetMyFeed = React.Dispatch<React.SetStateAction<MyFeedback[]>>;
+import { MyFeedback, SetMyFeed } from "model/userType";
 
 // 내 feedback GET요청
 const getMyFeed = async (setMyFeedback: SetMyFeed): Promise<void> => {
@@ -12,16 +8,17 @@ const getMyFeed = async (setMyFeedback: SetMyFeed): Promise<void> => {
   try {
     const response = await axios.get<MyFeedback[]>(uri);
     setMyFeedback(response.data);
-    console.log("response : ", response.data);
   } catch (error: unknown) {
-    if (axios.isAxiosError(error)) {  //axios error인지 확인
+    if (axios.isAxiosError(error)) {
+      //axios error인지 확인
       // 서버가 응답을 반환했지만 상태 코드가 오류를 나타내는 경우
       console.error(
         "Axios 오류 발생:",
         error.response?.status,
         error.response?.statusText
       );
-    } else if (error instanceof Error) {  //js오류인지 확인
+    } else if (error instanceof Error) {
+      //js오류인지 확인
       // 요청이 서버로 전송되었지만 응답을 받지 못한 경우
       console.error("일반 오류 발생:", error.message);
     } else {
