@@ -1,10 +1,14 @@
-import axios, {AxiosResponse} from "axios";
+import axios, { AxiosResponse } from "axios";
 
 // 댓글 달기 (post)
-const postFeed = async (answerId: number, comment: string, selected: number): Promise<AxiosResponse> => {
+const postFeed = async (
+  answerId: number,
+  comment: string,
+  selected: number
+): Promise<AxiosResponse> => {
   try {
     return await axios.post(`/api/comment/${answerId}`, {
-      comment,  // 댓글
+      comment, // 댓글
       commentPassFail: selected,
     });
   } catch (error: unknown) {
