@@ -1,5 +1,5 @@
 import axios from "axios";
-import { MemberRankList } from "model/Store";
+import { MemberRankList } from "model/userType";
 
 // 재학생 rank GET 요청
 const getJuniorRank = async (memberGen: number) => {
@@ -11,7 +11,10 @@ const getJuniorRank = async (memberGen: number) => {
       },
     };
 
-    const response = await axios.get<MemberRankList[]>(`/api/rank/junior`, config);
+    const response = await axios.get<MemberRankList[]>(
+      `/api/rank/junior`,
+      config
+    );
 
     if (response.status !== 204) {
       return response.data;
