@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback } from "react";
 import axios from "axios";
-import { MemberRankList } from "model/Store";
+import { MemberRankList } from "model/userType";
 
 // 재학생 등수 GET
 const GetSeniorsRank = () => {
@@ -13,7 +13,10 @@ const GetSeniorsRank = () => {
         withCredentials: true,
       };
 
-      const response = await axios.get<MemberRankList[]>(`/api/rank/senior`, config);
+      const response = await axios.get<MemberRankList[]>(
+        `/api/rank/senior`,
+        config
+      );
 
       if (response.status !== 204) setNewRank(response.data);
     } catch (error: unknown) {
