@@ -1,8 +1,8 @@
 import axios from "axios";
-import { SignInFormData } from "model/Store";
+import { SignInFormData } from "model/userType";
 
 const postLogin = async (formData: SignInFormData) => {
-  const uri = "api/signin";
+  const uri = "/api/signin";
 
   // 로그인 요청 api
   try {
