@@ -1,14 +1,9 @@
-import * as React from "react";
-import { ThemeProvider } from "styled-components";
-import { CssBaseline } from "@mui/material";
-import Tabs from "@mui/joy/Tabs";
-import Tab from "@mui/joy/Tab";
+import React from "react";
 import useMenu from "hooks/menu/useMenu";
-import {
-  StyledBox,
-  theme,
-  StyledTabList,
-} from "styles/styledComponent/Menu-styled";
+import { CssBaseline, Box, Tabs, Tab, createTheme, ThemeProvider } from "@mui/material";
+
+// MUI 테마 생성
+const theme = createTheme();
 
 const MenuBar: React.FC = () => {
   // 훅 사용
@@ -17,28 +12,45 @@ const MenuBar: React.FC = () => {
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
-      <StyledBox>
+      <Box sx={{ flexGrow: 1, mt: 2, overflow: "hidden" }}>
         <Tabs
           aria-label="Pipeline"
           value={index}
           onChange={(event, value) => {
             // value가 string | number | null이므로, 10진수로 변환
             if (value !== undefined && value !== null) {
-              const numericValue = typeof value === 'number' ? value : parseInt(value, 10);
+              const numericValue =
+                typeof value === "number" ? value : parseInt(value, 10);
               // number타입일때 탭이동
               if (!isNaN(numericValue)) {
                 toggleTab(numericValue);
               }
             }
           }}
+          sx={{ display: 'flex', justifyContent: 'flex-start', pt: 1 }}
         >
-          <StyledTabList index={index}>
-            <Tab value={1}>문제</Tab>
-            <Tab value={2}>공지사항</Tab>
-          </StyledTabList>
+          <Tab label="문제" value={1} sx={getTabStyle(index, 1)} />
+          <Tab label="공지사항" value={2} sx={getTabStyle(index, 2)} />
         </Tabs>
-      </StyledBox>
+      </Box>
     </ThemeProvider>
   );
-}
+};
+
+// Tab의 스타일 설정 함수
+const getTabStyle = (currentIndex: number | false, tabIndex: number) => ({
+  flex: 'initial',
+  fontSize: '26px',
+  color: currentIndex === tabIndex ? 'primary.main' : 'inherit',
+  '&::after': {
+    content: '""',
+    display: currentIndex === tabIndex ? 'block' : 'none',
+    height: '2.5px',
+    borderTopLeftRadius: '3px',
+    borderTopRightRadius: '3px',
+    backgroundColor: 'primary.main',
+    marginTop: '8px',
+  },
+});
+
 export default MenuBar;
