@@ -1,16 +1,25 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { StyledListItem } from "styles/styledComponent/ListItem-styled";
-import Button from "@mui/joy/Button";
-import { ProblemItemProps } from "model/Store";
+import { StyledListItem } from "styles/styledComponent";
+import { ProblemItemProps } from "model/problemType";
+// import Button from "@mui/material/Button";
+import { useSelector } from "react-redux";
+import { RootState } from "stores/store"; // RootState 임포트
 
-const ProblemItem: React.FC<ProblemItemProps> = (props) => {
+const ListItem: React.FC<ProblemItemProps> = (props) => {
   const navigate = useNavigate();
-  const memberStatus = sessionStorage.getItem("memberStatus");
+  const { memberStatus } = useSelector((state: RootState) => state.user); // redux에서 가져오기
 
+  // 해당 문제로 이동 메서드
   const goToProb = (answerId: number | null) => {
-    if (answerId) {
-      sessionStorage.setItem("problemId", props.pid.toString()); // sessionStorage저장은 문자열로 해야함
+    // problemId 저장
+    const problemId = props.pid.toString();
+    sessionStorage.setItem("problemId", problemId);
+
+    if(location.pathname === "/manageProb"){  // 관리자 페이지에서는 probDetail로 이동
+      navigate(`/manageProb/${problemId}`);
+    }
+    else if (answerId) {
       navigate(`/question/${props.answerId}`);
     } else {
       navigate(`/problem/${props.pid}`);
@@ -18,12 +27,13 @@ const ProblemItem: React.FC<ProblemItemProps> = (props) => {
   };
 
   //추가 by성임
-  // status가 "재학생"일 때는 props.state 값을 "pass"로 설정,
-  //그 외에는 props.state 값 그대로 사용
-  const state = memberStatus === "재학생" || "관리자" ? "success" : props.state;
+  // status가 "재학생"일 때는 props.state 값을 "success"로 설정,
+  // 그 외에는 props.state 값 그대로 사용
+  const state = memberStatus === "재학생" || memberStatus ==="관리자" ? "success" : props.state;
 
   return (
     <StyledListItem state={state}>
+      {/* 문제 클릭 시 페이지 이동*/}
       <div className="container">
         <div
           className="problem_id"
@@ -42,12 +52,17 @@ const ProblemItem: React.FC<ProblemItemProps> = (props) => {
           {props.ptitle}
         </div>
         <div className="grade">점수 {props.grade}</div>
-        <button
-          className="button"
-          onClick={() => navigate(`/solution/${props.pid}`)}
-        >
-          풀이보기
-        </button>
+
+        {/* 풀이보기 버튼: 관리자 페이지에서는 숨기기 */}
+        {location.pathname !== "/manageProb" && (
+          <button
+            className="button"
+            onClick={() => navigate(`/solution/${props.pid}`)}
+            disabled={state === null} // state가 null일 때 버튼 비활성화
+          >
+            풀이보기
+          </button>
+        )}
 
         {/* TODO : 비로그인 시 풀이보기 선택안되도록 */}
         {/* TODO : 상태에 따른 버튼UI 적용하기 */}
@@ -64,4 +79,4 @@ const ProblemItem: React.FC<ProblemItemProps> = (props) => {
   );
 };
 
-export default ProblemItem;
+export default ListItem;
