@@ -1,13 +1,8 @@
-import { useNavigate } from "react-router-dom";
-import "styles/css/manage.css";
-import { Member } from "model/Store";
+// 관리자_회원 컴포넌트
 import React from "react";
-
-// UserList 컴포넌트의 props 타입 정의
-interface UserListProps {
-  memberData: Member[];
-}
-
+import { useNavigate } from "react-router-dom";
+import "styles/css/manageUser.css";
+import { UserListProps } from "model/userType";
 // 회원목록 컴포넌트
 const UserList: React.FC<UserListProps> = ({ memberData }) => {
   const navigate = useNavigate();
@@ -40,7 +35,6 @@ const UserList: React.FC<UserListProps> = ({ memberData }) => {
           <li className="title_stunum">{mem.memberId}</li>
           <li className="title_studep">{mem.memberMajor}</li>
           <li className="title_none">
-            <button className="levelup_btn">등업</button>
             <button
               className="info_btn"
               onClick={() => gotoDetail(mem.memberId)}
