@@ -1,42 +1,50 @@
-import React from "react";
-import AceEditor from "react-ace";
-import "styles/css/App.css";
-import "ace-builds/src-noconflict/mode-c_cpp"; // C/C++ 언어 모드 import
-import "ace-builds/src-noconflict/theme-monokai"; // 다크 모드 테마 import
-
-interface CodeEditorProps{
-  code: string;  // 코드 내용
-  onChange: (text: string) => void;  // 코드 변경 핸들러
-  readOnly?: boolean;  // 읽기 전용 여부
-}
+import React, { useEffect, useRef } from "react";
+// import "styles/css/App.css";
+import { CodeEditorProps } from "model/problemType";
+import ace from "libs/ace";  // 커스텀 빌드된 ace.js 사용
+import "libs/mode-c_cpp"; // C/C++ 언어 모드 import
+import "libs/theme-monokai"; // 다크 모드 테마 import
+// import { ace as ace, mode_c_cpp, theme_monokai } from "libs"
 
 const CodeEditor: React.FC<CodeEditorProps> = ({ code, onChange, readOnly }) => {
-  // 부모 컴포넌트 코드 상태 업데이트
-  const handleChange = (newCode: string) => {
-    onChange(newCode);
-  };
+  const editorRef = useRef<HTMLDivElement | null>(null); // 커스텀빌드 라이브러리 사용위한 HOOK
 
-  return (
-    <AceEditor
-      mode="c_cpp"
-      theme="monokai"
-      onChange={handleChange} // handleChange 함수를 전달
-      readOnly={readOnly}
-      value={code} // 코드 표시를 위해 value props 사용
-      fontSize={16}
-      showPrintMargin={true}
-      showGutter={true}
-      highlightActiveLine={true}
-      setOptions={{
+  useEffect(() => {
+    if (editorRef.current) {
+      const editor = ace.edit(editorRef.current); // 'ace' 글로벌 객체로 에디터 초기화
+      editor.session.setMode("ace/mode/c_cpp"); // C/C++ 모드 설정
+      editor.setTheme("ace/theme/monokai"); // 모노카이 테마 설정
+      editor.setValue(code); // 초기 코드 설정
+      editor.setReadOnly(readOnly); // 읽기 전용 설정
+      editor.setFontSize(16); // 폰트 크기 설정
+
+      // 에디터 옵션 설정
+      editor.setOptions({
         enableBasicAutocompletion: true,
         enableLiveAutocompletion: true,
         enableSnippets: true,
         showLineNumbers: true,
         tabSize: 4,
-        cursorWidth: 2, // 커서 너비 설정
-      }}
-      style={{ width: "90%", height: "100%" }} // 필요에 따라 크기 조정
-      editorProps={{ $blockScrolling: true }}
+        cursorWidth: 2,
+      });
+
+      // 코드 변경 이벤트 처리
+      editor.on("change", () => {
+        const currentCode = editor.getValue();
+        onChange(currentCode);
+      });
+
+      // 컴포넌트가 언마운트될 때 에디터를 파괴
+      return () => {
+        editor.destroy();
+      };
+    }
+  }, [code, onChange, readOnly]);
+
+  return (
+    <div
+      ref={editorRef}
+      style={{ width: "90%", height: "100%" }} // 에디터 크기
     />
   );
 };
