@@ -1,16 +1,8 @@
-import React, { ChangeEvent } from "react";
-import CodeEditor from "components/problem/CodeEditor";
+import React, {Suspense } from "react";
 import { renderNewlines } from "components/common/Common";
-import { QuesAnswerType } from "model/Store";
-
-// ProbCode Props 타입 정의
-interface ProbCodeProps {
-  FstHandler: (e: ChangeEvent<HTMLTextAreaElement>) => void;
-  SecHandler: (e: ChangeEvent<HTMLTextAreaElement>) => void;
-  code: string;
-  qnA: QuesAnswerType;
-  setCode: (code: string) => void;
-}
+import { ProbCodeProps } from "model/questionType";
+import Loading from "components/common/Loading";
+const CodeEditor = React.lazy(() => import("components/problem/CodeEditor")); // Lazy Loading 적용
 
 // 작성했던 코드 출력 컴포넌트
 const ProbCode: React.FC<ProbCodeProps> = ({
@@ -19,17 +11,25 @@ const ProbCode: React.FC<ProbCodeProps> = ({
   code,
   qnA,
   setCode,
+  problemId
 }) => {
   return (
     <div className="code_question_container">
+
+      {/* 좌측 코드에디터 컴포넌트 */}
       <div className="code_container">
-        {/* 에디터로 수정_by성임 */}
-        <CodeEditor
-          onChange={setCode}
-          readOnly={true} // 코드 수정 비활성화
-          code={code}
-        ></CodeEditor>
+
+        {/* 에디터로 수정_by성임 - Lazy Loading 적용*/}
+        <Suspense fallback={<Loading/>}>
+          <CodeEditor
+            onChange={setCode}
+            readOnly={true}
+            code={code}
+          />
+        </Suspense>
       </div>
+
+      {/* 우측 질문답변 컴포넌트 */}
       <div className="question_container">
         {qnA.answerState === "question" ? (   // 기존 : !qnA.answerState || !qnA.answerSec
           // answerState: question (qnA만 있음, 답변해야 하는 상태)
@@ -51,6 +51,7 @@ const ProbCode: React.FC<ProbCodeProps> = ({
             </div>
           </>
         ) : (
+          // < answerState 참고 >
           // answerState: comment (qnA O / qnA.answerFst, qnA.answerSec O / feedback 0개 or 1개)
           // answerState: success
           // answerState: fail
