@@ -1,18 +1,13 @@
+import React, { useState } from "react";
 import useProbPage from "hooks/problem/useProbPage";
-import React, { useEffect, useState } from "react";
-import { StyledProblem } from "styles/styledComponent/Problem-styled";
 import { renderNewlines } from "components/common/Common";
+import { StyledProblem } from "styles/styledComponent/Problem-styled";
+import { ProbHeaderProps } from "model/problemType";
 
-// ProbHeader Props 타입
-interface ProbHeaderProps {
-  state: string;
-  navigate: (path: string) => void;
-}
+
 
 // 문제 헤더 컴포넌트
-const ProbHeader: React.FC<ProbHeaderProps> = ({ state, navigate }) => {
-  //const problemId: number = parseInt(sessionStorage.getItem("problemId") || '', 10); // 10진수로 변환해서 저장
-  const [problemId, setProblemId] = useState<number | null>(null);
+const ProbHeader: React.FC<ProbHeaderProps> = ({ state, navigate, problemId }) => {
   const [isDropdownOpen, setIsDropdownOpen] = useState<boolean>(false);
 
   // problem 제목, 내용 HOOK 호출 - null이나 undefined일때 -1전달
@@ -21,19 +16,7 @@ const ProbHeader: React.FC<ProbHeaderProps> = ({ state, navigate }) => {
   // 문제보기 dropdown 상태관리 함수
   const toggleDropdown = () => {
     setIsDropdownOpen((prevState ) => !prevState );
-  };
-
-  // 문제 ID 로딩
-  useEffect(() => {
-    const id = parseInt(sessionStorage.getItem("problemId") || '', 10);
-    // id가 NaN 아닌 경우
-    if (!isNaN(id)) {
-      setProblemId(id);
-    } else {
-      console.error("Invalid problem ID:", id);
-    }
-  }, []);
-
+  }
 
 
   // title과 problem이 null일 경우 기본값 설정 - TODO: 중복 수정하기
