@@ -6,23 +6,11 @@ const useMenu = () => {
   const location = useLocation();
   const navigate = useNavigate();
 
-  // 현재 경로에 따라 초기 탭 인덱스 설정
-  // const getInitialIndex = (): number | null => {
-  //   if (location.pathname === "/list")
-  //     return 1;
-  //   else if (location.pathname === "/notice") 
-  //     return 2;
-  //   else
-  //     return null; // 홈이나 기타 경로인 경우 탭이 선택되지 않도록
-  // };
-
-  // 탭 구분위한 index state
-  const [index, setIndex] = useState<number>(0);
+  const [index, setIndex] = useState<number | false>(false); // 탭 value값
 
   // 탭 이동
   const toggleTab = (value: number) => {
     setIndex(value);
-    console.log("value = ", value);
 
     if (value === 1) {
       navigate("/list");
@@ -31,16 +19,17 @@ const useMenu = () => {
     }
   };
 
+  // 탭 이동 훅
   useEffect(() => {
     // 경로가 '/'일 때 아무 탭도 선택되지 않도록 설정
     if (location.pathname === "/") {
-      setIndex(0);
+      setIndex(false);
     } else if (location.pathname === "/list") {
       setIndex(1);
     } else if (location.pathname === "/notice") {
       setIndex(2);
     } else {
-      setIndex(0); // 기타 경로일 경우 기본값으로 설정
+      setIndex(false); // 기타 경로일 경우 기본값으로 설정
     }
   }, [location.pathname, index]);
 
