@@ -1,44 +1,40 @@
 import { useState, useEffect } from "react";
-import getMyInfo from "apis/mypage/getMyInfo";
-import getMyProb from "apis/mypage/getMyProb";
-import getMyFeed from "apis/mypage/getMyFeed";
 import { useNavigate } from "react-router-dom";
-import { MemberDetail, Problem, MyFeedback } from "model/Store";
+import {getMyInfo, getMyProb, getMyFeed} from "apis/mypage";
+import { MyFeedback, MemberDetail, UseMypageReturnType } from "model/userType";
+import { Problem } from "model/problemType";
 
-// useMypage HOOK 반환 타입
-interface UseMypageReturnType {
-  memberData: MemberDetail | undefined;
-  myProb: Problem[];
-  myFeedback: MyFeedback[];
-}
-
-const useMypage = ():UseMypageReturnType => {
+// 마이페이지 api HOOK
+const useMypage = (status: boolean):UseMypageReturnType => {
   const navigate = useNavigate();
 
-  //정보리스트
+  // 내정보
   const [memberData, setMemberData] = useState<MemberDetail>();
-  //내문제리스트
+  // 내문제리스트
   const [myProb, setMyProb] = useState<Problem[]>([]);
-  //내피드백리스트
+  // 내피드백리스트
   const [myFeedback, setMyFeedback] = useState<MyFeedback[]>([]);
 
   // 관련 API요청 HOOK
   useEffect(() => {
-    getMyInfo(setMemberData); // 내 정보 GET 요청
-
     // 비로그인시
-    if (
-      sessionStorage.getItem("status") === null ||
-      sessionStorage.getItem("status") === undefined
-    ) {
-      console.log("로그인 안돼있음");
-
-      alert("로그인 해주세요.");
-      navigate("main");
+    if (status === null || status === undefined) {
+      alert("로그인 해주세요");
+      navigate("/signin");
       return;
     }
-    getMyProb(setMyProb); // 내가 푼 문제 GET요청
-    getMyFeed(setMyFeedback); // 내 feedback GET요청
+
+    // 로그인 상태일 때 API 요청
+    const fetchData = async() => {
+      try{
+        getMyInfo(setMemberData); // 내 정보 GET요청
+        getMyProb(setMyProb); // 내가 푼 문제 GET요청
+        getMyFeed(setMyFeedback); // 내 feedback GET요청
+      }catch(error){
+        console.error("데이터를 가져오는 중 오류 발생:", error);
+      }
+    }
+    fetchData();
   }, [navigate]); // navigate가 최신상태인지 확인
 
   return { memberData, myProb, myFeedback };
