@@ -1,11 +1,11 @@
-import getJuniorRank from "apis/ranking/getJuniorRank";
 import { useState, useEffect, useCallback } from "react";
-import { MemberRankList } from "model/Store";
+import getJuniorRank from "apis/ranking/getJuniorRank";
+import { MemberRankList } from "model/userType";
 import { SelectChangeEvent } from "@mui/material";
 
 // 신입생 등수 GET
 const useJuniorRank = () => {
-  const [memberGen, setMemberGen] = useState<number>(35); // memberGen 35기로 기본 설정
+  const [memberGen, setMemberGen] = useState<number>(36); // memberGen 35기로 기본 설정
   const [newRank, setNewRank] = useState<MemberRankList[]>([]);
 
   const handleSelectChange = (e : SelectChangeEvent<string>) => {
